Network Scanner/requirements.txt
python-nmap
aiohttp
scapy
asyncio